% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anr}
\alias{anr}
\title{Runs Expectation Maximization ANR and reestimates model parameters.}
\usage{
anr(fasta, alpha, beta, cutoff, niter, w = 0.5)
}
\arguments{
\item{fasta}{Dataset of sequences.}

\item{alpha}{PWM model.}

\item{beta}{0-order Markov Chain.}

\item{cutoff}{Cutoff for EM convergence.}

\item{niter}{Maximum number of iterations.}

\item{w}{Priori probability for motif belongs to a positive model.}
}
\value{
Updated PWM model.
}
\description{
Runs Expectation Maximization ANR and reestimates model parameters.
}
